loads of introspection to make validation work
I've gone full hog and abstracted out the fields to a Field class, which has subclasses representing different types of data.
The fields are defined as members of the Notebook class, much like django fields.

A field loads data from the path it's given, then cleans it to obtain a finished value that can be used.
If the data is invalid, the `clean` method should raise it.

A notebook is valid if all its fields are valid. Notebook.is_valid will raise a NotebookInvalidException with all of the field errors if any are invalid.